+++ src/main/kotlin/org/sert2521/infiniterecharge2020/OI.kt
@@ -1,0 +1,22 @@
+package org.sert2521.infiniterecharge2020
+
+import edu.wpi.first.wpilibj.Joystick
+import edu.wpi.first.wpilibj.XboxController
+import edu.wpi.first.wpilibj.smartdashboard.SendableChooser
+
+object OI {
+    enum class ControlMode {
+        CONTROLLER, JOYSTICK
+    }
+
+    val controlModeChooser = SendableChooser<ControlMode>().apply {
+        addOption("Joystick", ControlMode.JOYSTICK)
+        addOption("Controller", ControlMode.CONTROLLER)
+    }
+
+    val controlMode get() = controlModeChooser.selected ?: ControlMode.JOYSTICK
+
+    val primaryController by lazy { XboxController(Operator.PRIMARY_CONTROLLER) }
+    val primaryJoystick by lazy { Joystick(Operator.PRIMARY_STICK) }
+    val secondaryJoystick by lazy { Joystick(Operator.SECONDARY_STICK) }
+}
